add new psd functions and flexibility
